requirements.txt
Flask>=2.3
openai>=1.0.0
python-dotenv>=1.0